System Builder Database Connection Part 2
System Builder Database Connection Part 2,
some adjustments made on CreatinScript.sql

+++ Backend/PC-Part-Picker/Models/M2.cs
@@ -6,9 +6,9 @@ public class M2
 	public string Name { get; set; }
 	public string Manufacturer { get; set; }
 	public byte[]? Image { get; set; }
-	public float Price { get; set; }
+	public double Price { get; set; }
 	public int Volume { get; set; }
 	public int Read_Speed { get; set; }
 	public int Write_Speed { get; set; }
-	public float PCIe_Version { get; set; }
+	public double PCIe_Version { get; set; }
 }
